分类，员工完成

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,47 +2,87 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
+import com.sky.constant.StatusConstant;
+import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.dto.PageResult;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.service.CategoryService;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     @Override
     public PageResult categoryPage(CategoryPageQueryDTO page) {
         PageHelper.startPage(page.getPage(), page.getPageSize());
-        Page<Category> category = (Page<Category>) categoryMapper.selectPage(page.getName(), page.getType());
-        return new PageResult(category.getTotal(),category.getResult());
+        Page<Category> category = categoryMapper.selectPage(page.getName(), page.getType());
+        return new PageResult(category.getTotal(), category.getResult());
     }
 
     @Override
-    public boolean addCategory(CategoryDTO category) {
+    public boolean addCategory(CategoryDTO categoryDTO) {
+        Category category = new Category();
+        BeanUtils.copyProperties(categoryDTO, category);
+        category.setStatus(StatusConstant.DISABLE);
+        category.setCreateTime(LocalDateTime.now());
+        category.setUpdateTime(LocalDateTime.now());
+        category.setCreateUser(BaseContext.getCurrentId());
+        category.setUpdateUser(BaseContext.getCurrentId());
         return categoryMapper.insertCategory(category) == 1;
     }
 
     @Override
-    public void changeStatus(Integer status , Long id) {
-        Category category = new Category();
-        category.setStatus(status);
-        category.setId(id);
+    public void changeStatus(Integer status, Long id) {
+        Category category = Category.builder()
+                .id(id)
+                .status(status)
+                .updateTime(LocalDateTime.now())
+                .updateUser(BaseContext.getCurrentId())
+                .build();
         categoryMapper.updateCategory(category);
     }
 
     @Override
-    public void updateEmp(Category category) {
+    public void updateEmp(CategoryDTO categoryDTO) {
+        Category category = new Category();
+        BeanUtils.copyProperties(categoryDTO, category);
+        //设置修改时间、修改人
+        category.setUpdateTime(LocalDateTime.now());
+        category.setUpdateUser(BaseContext.getCurrentId());
         categoryMapper.updateCategory(category);
     }
 
     @Override
     public void deleteEmp(Long id) {
+        //查询当前类是否有关联菜品 如果有则抛出异常
+        int count = dishMapper.countByCategoryId(id);
+        if (count > 0) {
+            //当前分类下有菜品，不能删除
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        //查询当前分类是否关联了套餐，如果关联了就抛出业务异常
+        count = setmealMapper.countByCategoryId(id);
+        if (count > 0) {
+            //当前分类下有菜品，不能删除
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.deleteEmp(id);
     }
 }
